bug fixd
